+++ src/shared/libs/cookie/deleteCookies.ts
