+++ src/views/form/edit/api/getEditSurveyForm.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientInstance from '@/shared/libs/clientInstance';
+import clientInstance from '@/shared/libs/http/clientInstance';
 
 export const getEditSurveyForm = async (id: string, type: string) => {
   try {
