+++ src/widgets/SMS/api/sendSMS.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 import { SendSmSData } from '@/shared/types/sms';
 
 export const sendSMS = async (
