+++ src/widgets/admin/api/deleteExpoApi.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 
 export const deleteExpoApi = async (id: number): Promise<void> => {
   try {
