+++ src/widgets/admin/api/getAdminData.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 import { AdminData } from '@/shared/types/admin/type';
 
 export const getAdminData = async (): Promise<AdminData> => {
