+++ src/widgets/application/api/getApplicationForm.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientInstance from '@/shared/libs/clientInstance';
+import clientInstance from '@/shared/libs/http/clientInstance';
 
 export const getApplicationForm = async (id: string, userType: string) => {
   try {
