+++ src/widgets/exhibition/api/uploadImage.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 
 export const uploadImage = async (file: File | null | string) => {
   if (!file) return null;
