+++ src/widgets/exhibition/edit/api/editExhibition.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 import { EditExhibitionData } from '@/shared/types/exhibition/edit/type';
 
 export const editExhibition = async (data: EditExhibitionData, id: number) => {
