+++ src/widgets/name-tag/api/patchAttendUser.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 import {
   AttendUserQrRequest,
   AttendUserResponse,
