+++ src/widgets/program-detail/api/patchStandardAttendance.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 import { PatchStandardProgramData } from '@/shared/types/program-detail/type';
 
 export const patchStandardAttendance = async ({
