+++ src/widgets/program-detail/api/patchTrainingAttendance.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import clientTokenInstance from '@/shared/libs/clientTokenInstance';
+import clientTokenInstance from '@/shared/libs/http/clientTokenInstance';
 import { PatchTrainingProgramData } from '@/shared/types/program-detail/type';
 
 export const patchTrainingAttendance = async ({
